requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
ffmpeg-python==0.2.0
future==0.18.3
idna==3.4
pytube==15.0.0
requests==2.31.0
urllib3==2.0.7
